adding primeChecker.java
